pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["pytest", "pytest-asyncio"]

[project]
name = "najs"
description = "Simple helper functions for utilizing nats-py"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = ["async-lru", "fastavro", "loguru", "nats-py", "nkeys", "tenacity"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
src = ["src"]
exclude = [
  # Don't want to fix format on NATSD which is copied from nats-py
  "tests/conftest.py",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "D", # pydocstyle, we do not enforce this
]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false

[tool.ruff.lint.per-file-ignores]
"**/tests/**/test_*.py" = [
  "S101", # asserts allowed in tests...
  "ANN", # Annotation
  "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
  "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
  "PLR2004", # Magic value used in comparison, ...
  "INP001", # Files in test are not part of package
  "PT004", # Leading underscore on functions that dont return anything
  "SLF001", # Allow access private class members
]

[tool.uv.extra-build-dependencies]
nkeys = ["setuptools"]
